HAAR-125 Move calls to auth/user/me now redirected via manage-users-api (#461)
