+++ backend/middleware/currentUser.test.js
@@ -1,18 +1,17 @@
 const currentUser = require('./currentUser')
 
 describe('Current user', () => {
-  const oauthApi = {}
   const nomisUsersAndRolesApi = {}
   const manageUsersApi = {}
   let req
   let res
 
   beforeEach(() => {
-    oauthApi.currentUser = jest.fn()
+    manageUsersApi.currentUser = jest.fn()
     manageUsersApi.currentRoles = jest.fn()
     nomisUsersAndRolesApi.currentUserCaseloads = jest.fn()
 
-    oauthApi.currentUser.mockReturnValue({ name: 'Bob Smith', activeCaseLoadId: 'MDI' })
+    manageUsersApi.currentUser.mockReturnValue({ name: 'Bob Smith', activeCaseLoadId: 'MDI' })
     manageUsersApi.currentRoles.mockReturnValue([{ roleCode: 'FRED' }])
     nomisUsersAndRolesApi.currentUserCaseloads.mockReturnValue({
       username: 'user',
@@ -28,10 +27,10 @@ describe('Current user', () => {
   })
 
   it('should request and store user details', async () => {
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
     await controller(req, res, () => {})
 
-    expect(oauthApi.currentUser).toHaveBeenCalled()
+    expect(manageUsersApi.currentUser).toHaveBeenCalled()
     expect(manageUsersApi.currentRoles).toHaveBeenCalled()
     expect(req.session.userDetails).toEqual({ name: 'Bob Smith', activeCaseLoadId: 'MDI' })
     expect(req.session.userRoles).toEqual({
@@ -48,7 +47,7 @@ describe('Current user', () => {
 
   it('should stash data into res.locals', async () => {
     req.get.mockReturnValue('host')
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
 
     await controller(req, res, () => {})
     expect(res.locals.user).toEqual({
@@ -82,7 +81,7 @@ describe('Current user', () => {
 
   it('should set group manager', async () => {
     manageUsersApi.currentRoles.mockReturnValue([{ roleCode: 'FRED' }, { roleCode: 'AUTH_GROUP_MANAGER' }])
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
 
     await controller(req, res, () => {})
 
@@ -100,7 +99,7 @@ describe('Current user', () => {
 
   it('should set maintain access', async () => {
     manageUsersApi.currentRoles.mockReturnValue([{ roleCode: 'FRED' }, { roleCode: 'MAINTAIN_ACCESS_ROLES' }])
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
 
     await controller(req, res, () => {})
 
@@ -118,7 +117,7 @@ describe('Current user', () => {
 
   it('should set maintain access admin', async () => {
     manageUsersApi.currentRoles.mockReturnValue([{ roleCode: 'FRED' }, { roleCode: 'MAINTAIN_ACCESS_ROLES_ADMIN' }])
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
 
     await controller(req, res, () => {})
 
@@ -136,7 +135,7 @@ describe('Current user', () => {
 
   it('should set Search for an external user', async () => {
     manageUsersApi.currentRoles.mockReturnValue([{ roleCode: 'FRED' }, { roleCode: 'MAINTAIN_OAUTH_USERS' }])
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
 
     await controller(req, res, () => {})
 
@@ -154,7 +153,7 @@ describe('Current user', () => {
 
   it('should set create dps user', async () => {
     manageUsersApi.currentRoles.mockReturnValue([{ roleCode: 'FRED' }, { roleCode: 'CREATE_USER' }])
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
 
     await controller(req, res, () => {})
 
@@ -171,7 +170,7 @@ describe('Current user', () => {
   })
   it('should set maintain dps user', async () => {
     manageUsersApi.currentRoles.mockReturnValue([{ roleCode: 'FRED' }, { roleCode: 'MANAGE_NOMIS_USER_ACCOUNT' }])
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
 
     await controller(req, res, () => {})
 
@@ -189,7 +188,7 @@ describe('Current user', () => {
 
   it('should set Role admin for an external user', async () => {
     manageUsersApi.currentRoles.mockReturnValue([{ roleCode: 'FRED' }, { roleCode: 'ROLES_ADMIN' }])
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
 
     await controller(req, res, () => {})
 
@@ -207,7 +206,7 @@ describe('Current user', () => {
 
   it('should set Role manage email domains for an external user', async () => {
     manageUsersApi.currentRoles.mockReturnValue([{ roleCode: 'FRED' }, { roleCode: 'MAINTAIN_EMAIL_DOMAINS' }])
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
 
     await controller(req, res, () => {})
 
@@ -235,7 +234,7 @@ describe('Current user', () => {
       { roleCode: 'MANAGE_NOMIS_USER_ACCOUNT' },
       { roleCode: 'MAINTAIN_EMAIL_DOMAINS' },
     ])
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
 
     await controller(req, res, () => {})
 
@@ -252,14 +251,14 @@ describe('Current user', () => {
   })
 
   it('ignore xhr requests', async () => {
-    const controller = currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi })
+    const controller = currentUser({ nomisUsersAndRolesApi, manageUsersApi })
     req.xhr = true
 
     const next = jest.fn()
 
     await controller(req, res, next)
 
-    expect(oauthApi.currentUser.mock.calls.length).toEqual(0)
+    expect(manageUsersApi.currentUser.mock.calls.length).toEqual(0)
     expect(manageUsersApi.currentRoles.mock.calls.length).toEqual(0)
     expect(next).toHaveBeenCalled()
   })
