+++ backend/routes.js
@@ -20,7 +20,7 @@ const config = require('./config').default
 const configureRoutes = ({ oauthApi, manageUsersApi, nomisUsersAndRolesApi }) => {
   const router = express.Router()
 
-  router.use(currentUser({ oauthApi, nomisUsersAndRolesApi, manageUsersApi }))
+  router.use(currentUser({ nomisUsersAndRolesApi, manageUsersApi }))
   router.use(featureSwitches(config))
 
   router.use('/', menuRouter({ manageUsersApi }))
